chore: remove legacy Pool and exchange
