+++ src/pools/index.ts
@@ -1,0 +1,7 @@
+import { getPool } from "./poolConstructor";
+import { PoolTemplate } from "./PoolTemplate";
+
+export {
+    getPool,
+    PoolTemplate,
+};
