+++ src/utils.ts
@@ -389,12 +389,6 @@ export const _getFactoryStatsPolygon = memoize(
     }
 )
 
-export const getPoolList = (): string[] => Object.keys(curve.constants.POOLS_DATA);
-
-export const getFactoryPoolList = (): string[] => Object.keys(curve.constants.FACTORY_POOLS_DATA);
-
-export const getCryptoFactoryPoolList = (): string[] => Object.keys(curve.constants.CRYPTO_FACTORY_POOLS_DATA);
-
 export const getUsdRate = async (coin: string): Promise<number> => {
     const [coinAddress] = _getCoinAddresses(coin);
     return await _getUsdRate(coinAddress);
